requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
blinker==1.9.0
boto3==1.42.22
botocore==1.42.22
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.1.0
click==8.1.8
commonmark==0.9.1
cryptography==43.0.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
envs==1.4
filelock==3.20.3
Flask==3.1.2
frozenlist==1.8.0
future==0.18.3
ggshield==1.47.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
MarkupSafe==3.0.3
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
multidict==6.7.1
mypy_extensions==1.1.0
nodeenv==1.10.0
oauth==1.0.1
oauthlib==3.2.2
packaging==25.0
platformdirs==3.0.0
pre_commit==4.5.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.3.1
pydantic==2.12.5
pydantic_core==2.41.5
pygitguardian==1.28.0
Pygments==2.19.2
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-jose-cryptodome==1.3.2
PyYAML==6.0.3
requests==2.32.5
rich==12.5.1
s3transfer==0.16.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
truststore==0.10.4
twilio==9.10.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.36.1
warrant==0.6.1
Werkzeug==3.1.4
yarl==1.22.0
